+++ src/components/header/components/SearchBox.tsx
@@ -1,12 +1,13 @@
+import { Cancel, Search } from 'iconoir-react';
 import * as _ from 'lodash';
-import { useCallback, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { useIntl } from 'react-intl';
-import { Input } from '../../../ui/Input';
-import SearchResult from './SearchResult';
-import { Search } from 'iconoir-react';
-import { Spin } from '../../../ui/Spin';
-import { Flex } from '../../../ui/Flex';
+import { ActionButton } from '../../../ui/ActionButton';
 import { styled } from '../../../ui/design-system';
+import { Flex } from '../../../ui/Flex';
+import { Input } from '../../../ui/Input';
+import { Spin } from '../../../ui/Spin';
+import SearchResult from './SearchResults';
 
 const StyledSearchResult = styled("div", {
   position: "absolute",
@@ -27,6 +28,14 @@ const SearchBox = () => {
   const [searchText, setSearchText] = useState('');
   const formatMessage = intl.formatMessage;
   const [loadingSearchResult, setLoadingSearchResult] = useState(false);
+  const inputEl = useRef(null);
+
+  const clearText = () => {
+    setSearchText("");
+    if (inputEl.current) {
+      inputEl.current.value = "";
+    }
+  }
 
   const debouncedSetSearchText = useCallback(
     _.debounce(
@@ -39,32 +48,47 @@ const SearchBox = () => {
     []
   );
 
+  let endSlot = null;
+
+  if (loadingSearchResult) {
+    endSlot = <Spin size="sm" />
+  }
+  else if (searchText) {
+    endSlot = <ActionButton>
+      <Cancel onClick={() => {
+        clearText();
+      }} />
+    </ActionButton>
+  }
+
   const searchBox = (
-    <Flex align="both" css={{margin: "0 $12", flex: 1, position: "relative"}}>
+    <Flex align="both" css={{ margin: "0 $8", flex: 1, position: "relative" }}
+      onBlur={(e) => {
+        // DO NOT MODIFY THIS METHOD
+        // Ref: https://gist.github.com/pstoica/4323d3e6e37e8a23dd59
+        const currentTarget = e.currentTarget;
+        // Check the newly focused element in the next tick of the event loop
+        setTimeout(() => {
+          // Check if the new activeElement is a child of the original container
+          if (!currentTarget.contains(document.activeElement)) {
+            // You can invoke a callback or add custom logic here
+            clearText();
+          }
+        }, 0);
+      }}
+    >
       <Input
         placeholder={formatMessage({ id: 'app.header.search.box' })}
         startSlot={<Search />}
-        endSlot={loadingSearchResult ? <Spin size="sm" /> : null}
+        endSlot={endSlot}
         size="md"
+        ref={inputEl}
         onChange={(e) => {
           debouncedSetSearchText(e.target.value);
         }}
       ></Input>
       {searchText ? (
         <StyledSearchResult
-          onBlur={(e) => {
-            // DO NOT MODIFY THIS METHOD
-            // Ref: https://gist.github.com/pstoica/4323d3e6e37e8a23dd59
-            const currentTarget = e.currentTarget;
-            // Check the newly focused element in the next tick of the event loop
-            setTimeout(() => {
-              // Check if the new activeElement is a child of the original container
-              if (!currentTarget.contains(document.activeElement)) {
-                // You can invoke a callback or add custom logic here
-                setSearchText('');
-              }
-            }, 0);
-          }}
         >
           <SearchResult {...{ searchText, setLoadingSearchResult }}></SearchResult>
         </StyledSearchResult>
